src/mcp_doc_retriever/context7/pdf_extractor/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
sse-starlette>=1.8.2
python-multipart>=0.0.9
pymupdf>=1.24.10
camelot-py[cv]>=0.11.0
pdf2image>=1.17.0
pytesseract>=0.3.13
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.1
cleantext>=1.1.4
tiktoken>=0.7.0
torch>=2.4.1
transformers>=4.44.2
markdown-it-py>=3.0.0
loguru>=0.7.2
marker-pdf>=0.2.17
gunicorn>=21.2.0